Implement unittest
Created a 404 unittest for pagination that tests request for a page
out of range for a proper 404 response.

Modifited the unittest for get_pages to verify the length of the
questions matches the QUESTIONS_PER_PAGE configuration. To do this
I moved the QUESTIONS_PER_PAGE constant into a new config.py file and
imported into both __init__.py and test_flaskr.py.

Removed the .limit(10) from the query in the '/api/questions' route
handler, and verified that the test for get_question now fails when
more than QUESTIONS_PER_PAGE is in the response.