Implement pagination, fix error handling, unittest
Implemented pagination for get_questions.

Using Exceptions meant that aborts raised exceptions,
and the default exception of abort(422) was returning in
all cases (e.g. abort(404) still returned status code 422).
Solution for now is to manually raise exceptions with the
status code as a string, and use conditionals in 'except':
to abort appropriately.

Added additional checks to unittest for the get_questions
route to verify that 'total_questions', 'current_category'
and 'categories' are validated.