Implement unittests for adding questions
Implemented a unittest to verify new questions can be added using
the post method on the /api/questions route and implemented a unittest
to verify that submitting empty question fields will return a 422
error.

Also: removed 405 unittests for POST and DELETE methods on the
/api/questions route, while adding 405 unittests for PUT and PATCH
methods on the /api/categories route and the /api/questions routes.