Post question validation and unnittests
Added validation to the PostQuestion class that verifies difficulties
and categories are passed as integers. Validation also verifies that
a category exists in the database and that the difficulty is in the
range of one to five.

Created unnittests to verify that these validators work, and that all
previous functionality works.